Move and rename many more os files
