+++ lib/src/memzero.c
