+++ lib/src/os/osViSetMode.c
