+++ src/main.c
@@ -4,7 +4,7 @@
 #include "main.h"
 #include "thread0_epc.h"
 #include "unknown_06B2B0.h"
-#include "lib/src/xprintf.h"
+#include "lib/src/libc/xprintf.h"
 
 /************ .rodata ************/
