+++ src/screen_asset.c
@@ -2,7 +2,7 @@
 /* RAM_POS: 0x8007F640 */
 
 #include "screen_asset.h"
-#include "lib/src/xprintf.h"
+#include "lib/src/libc/xprintf.h"
 
 // Render settings for the screens
 Gfx gRdpSetModeScreenAsset[] = {
